In the convenience file_update structure, we sometimes have the wantmd5 field point at memory directly in the read buffer, in which case we must not try to free() it. This is most likely a petty micro-optimization, but nevertheless, it was achieved by setting this field to NULL after we were done with it (so, in a place that was potentially far from the initial assignment).
This has confused myself and others in the past, so try to sort this out once
and for all. Have a free_wantmd5 field in the file_update structure indicating
whether the wantmd5 field needs to be freed or not. In the few places where we
want that behaviour, it can be set right after the field has been initialized,
making the whole thing a lot more obvious.

Also rename fup_cleanup() to fup_reset(), I find it more intuitive.